+++ StarCitizenApi/ArkStarmap/Model/CelestialObjects/Shader_Data1.cs
@@ -5,18 +5,18 @@ namespace StarCitizenApi.ArkStarmap.Model.CelestialObjects
     public class Shader_Data1
     {
         [JsonProperty("orbitalMin")]
-        public int orbitalMin { get; set; }
+        public double orbitalMin { get; set; }
 
         [JsonProperty("orbitalMax")]
-        public int orbitalMax { get; set; }
+        public double orbitalMax { get; set; }
 
         [JsonProperty("orbitalFactor")]
-        public int orbitalFactor { get; set; }
+        public double orbitalFactor { get; set; }
 
         [JsonProperty("highlight")]
         public Highlight1 highlight { get; set; }
 
         [JsonProperty("radius")]
-        public float radius { get; set; }
+        public double radius { get; set; }
     }
 }
